+++ casper/src/test/scala/coop/rchain/casper/util/comm/CasperPacketHandlerSpec.scala
@@ -131,77 +131,6 @@ class CasperPacketHandlerSpec extends WordSpec {
 
   "CasperPacketHandler" when {
 
-    "in  BootstrapCasperHandler state" should {
-      "make a transition to ApprovedBlockReceivedHandler once ApprovedBlock has been received" in {
-        import monix.execution.Scheduler.Implicits.global
-        val fixture = setup()
-        import fixture._
-
-        val validators = Set(ByteString.copyFrom(validatorPk.bytes))
-
-        val theInit = Task.unit
-
-        implicit val engineCell = Cell.unsafe[Task, Engine[Task]](Engine.noop)
-
-        // interval and duration don't really matter since we don't require and signs from validators
-        val bootstrapCasper =
-          new BootstrapCasperHandler[Task](
-            runtimeManager,
-            shardId,
-            Some(validatorId),
-            validators,
-            theInit
-          )
-
-        val approvedBlockCandidate = ApprovedBlockCandidate(block = Some(genesis))
-
-        val approvedBlock: ApprovedBlock = ApprovedBlock(
-          candidate = Some(approvedBlockCandidate),
-          sigs = Seq(
-            Signature(
-              ByteString.copyFrom(validatorPk.bytes),
-              "ed25519",
-              ByteString.copyFrom(
-                Ed25519.sign(Blake2b256.hash(approvedBlockCandidate.toByteArray), validatorSk)
-              )
-            )
-          )
-        )
-
-        val approvedPacket = Packet(transport.ApprovedBlock.id, approvedBlock.toByteString)
-
-        val test = for {
-          _                   <- EngineCell[Task].set(bootstrapCasper)
-          casperPacketHandler = new CasperPacketHandler[Task]
-          _                   <- casperPacketHandler.handle(local).apply(approvedPacket)
-          casperO             <- MultiParentCasperRef[Task].get
-          _                   = assert(casperO.isDefined)
-          blockO              <- blockStore.get(genesis.blockHash)
-          _                   = assert(blockO.isDefined)
-          _                   = assert(blockO.contains(genesis))
-          handlerInternal     <- EngineCell[Task].read
-          _                   = assert(handlerInternal.isInstanceOf[ApprovedBlockReceivedHandler[Task]])
-          _ = assert(
-            transportLayer.requests.head.msg == packet(
-              local,
-              networkId,
-              transport.ForkChoiceTipRequest,
-              ByteString.EMPTY
-            )
-          )
-          _ = transportLayer.reset()
-          // assert that we really serve last approved block
-          lastApprovedBlockO <- LastApprovedBlock[Task].get
-          _                  = assert(lastApprovedBlockO.isDefined)
-          _                  <- casperPacketHandler.handle(local)(approvedBlockRequestPacket)
-          head               = transportLayer.requests.head
-          _                  = assert(head.msg.message.packet.get.content == approvedBlock.toByteString)
-        } yield ()
-
-        test.unsafeRunSync
-      }
-    }
-
     "in ApprovedBlockReceivedHandler state" should {
       import monix.execution.Scheduler.Implicits.global
       val fixture = setup()
